Move execSpecial Ops into their own file
